Change in division method

+++ src/main/java/co/com/sofka/app/calculator/BasicCalculator.java
@@ -22,7 +22,7 @@ public class BasicCalculator {
         return number1 * number2;
     }
 
-    public float division(Long number1, Long number2) {
+    public long division(Long number1, Long number2) {
         logger.info( "Summing {} / {}", number1, number2 );
         try{
